deepfake-backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9          # required for UploadFile / Form

# ── ML stack ─────────────────────────────────────────────────────────────────
torch==2.2.2                    # or torch==2.3.1+cu121 for CUDA 12.1
torchvision==0.17.2
numpy==1.26.4
Pillow==10.2.0
opencv-python-headless==4.10.0.84

# ── Face detection (optional but strongly recommended) ────────────────────────
facenet-pytorch==2.6.0           # MTCNN — remove if causing issues, Haar fallback activates

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest==8.3.2
httpx==0.27.2                    # TestClient transport
pytest-asyncio==0.23.8
